Remove reference to Plug.Conn struct in Phoenix.EventHandler
Instead of requiring a Plug.Conn, this patch changes the match
in Phoenix.EventHandler.handle_event/4 to accept any map with a :conn
that has a :private key.

    %{conn: %{private: %{appsignal_transaction: transaction}}},

This removes the hard dependency on Plug, allowing this module to be
compiled in applications that don't use Phoenix or Plug without having
to add a hard dependency on the latter.